Restructure for CentOS 7 RPM
We moved off of a Bashton CentOS 6 AMI to the official CentOS 7 AMI,
and didn't realize that the Bashton AMI was providing a version of
ec2-net-utils. Without this the changes made by keepalived when failing
over servers does not work because the API calls moving the network
interfaces and several local IPs is won't do anything without the
required network config changes.